api/pyproject.toml
[project]
name = "api"
version = "0.1.0"
description = "Medical report simplification API with validation pipeline"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    # Core framework
    "fastapi[standard]>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "python-multipart>=0.0.6",
    # Validation & types
    "pydantic>=2.0.0",
    # AI/ML
    "google-genai>=1.60.0",
    "openai>=1.0.0",
    "sentence-transformers>=5.2.0",
    "transformers>=4.36.0",
    "torch>=2.0.0",
    "scikit-learn>=1.3.0",
    # Medical NLP
    "medspacy>=1.3.1",
    # Search & data
    "opensearch-py>=2.4.0",
    # PDF processing
    "PyPDF2>=3.0.0",
    "pypdf>=6.6.0",
    "pymupdf>=1.24.0",
    # External APIs
    "exa-py>=2.1.1",
    # Environment
    "python-dotenv>=1.0.0",
    "spacy>=3.8.11",
    "typer>=0.21.1",
    "pip>=25.3",
    "en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl",
    "mlflow>=1.27.0",
    "bert-score>=0.3.13",
    "pandas==3.0.0",
    "scispacy>=0.2.4",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "httpx>=0.25.0",
]

[project.scripts]
run-evals = "scripts.run_evals:app"
